% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggR_possible.R
\name{aggR_possible}
\alias{aggR_possible}
\title{aggR_possible}
\usage{
aggR_possible(data, number = NULL, features = names(data), keep = TRUE,
  all_symbol = "")
}
\arguments{
\item{data}{data.frame}

\item{number}{integer}

\item{features}{vector}

\item{keep}{logical}

\item{all_symbol}{character}
}
\value{
filter_list list
}
\description{
Checks through a dataframe and a vector of features that can be deleted
OR KEPT!!!!!
outputs a list of list
each sublist contains a set of features that can be filtered out.
To filter by keeping, set keep = TRUE (default)
To filter by columns to filter out, set keep = FALSE
if you want to ignore certain columns, use the features setting.
If you do, the function will ONLY consider these columns, ignoring the others.
If you only want to look at combinations of a certain length, use number.  Otherwise, leave it as null
}
\examples{
load("../example_data/yummy.Rda")
dat_2

aggR_possible(dat_2, features =c("Colour", "Sweet_or_Salty", "Fruit"), keep = TRUE)
aggR_possible(dat_2, features =c("Colour", "Sweet_or_Salty", "Fruit"), number = 2, keep = TRUE)


}
