% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRviz_filter.R
\name{aggRviz_filter}
\alias{aggRviz_filter}
\title{Filter aggregated data to a chosen level}
\usage{
aggRviz_filter(dat, col_2_delete)
}
\arguments{
\item{dat}{data.frame}

\item{col_2_delete}{vector}
}
\value{
data.frame
}
\description{
The columns to delete are the features that you DO NOT want to stratify by.
This function filters out any row, stratified by those columns,
and filters out any unstratified rows from the other features.
}
\examples{


}
