% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_blanks.R
\name{filter_blanks}
\alias{filter_blanks}
\title{Filter aggregated data to the most granular level}
\usage{
filter_blanks(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Create a filtered dataframe with no blanks
}
\examples{
filter_blanks(my_df)

}
