% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_measures.R
\name{identify_measures}
\alias{identify_measures}
\title{identify_measures}
\usage{
identify_measures(data, key = c("measure", "rate"))
}
\arguments{
\item{data}{data.frame or vector}

\item{key}{vector}
}
\value{
vector
}
\description{
Enter a data.frame or vector, and this function will return a vector of measures
A measure is defined as a column name (or item in a vector) that contains a key word.
default is c("measure", "rate").
The function uses stringr to detect these regular expressions.
}
\examples{
load("../example_data/yummy.Rda")
names(dat_1)
identify_measures(dat_1)

names_to_check <- c("condition", "odds_ratio", "blood_type", "percent_ratio" )
identify_measures(names_to_check, key = "ratio")

}
