% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_set.R
\name{join_set}
\alias{join_set}
\title{Join 2 tables on the maximum overlapping set of features}
\usage{
join_set(x, y)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}
}
\value{
data.frame
}
\description{
This function will pick key column (such as "interesting rate" or proportion rate)
and join two lowest datasets by common columns.
}
\examples{
join_set(A,B)
A is a dataset with feature "a","b","proportion rate"
B is a dataset with feature "a","c","interesting rate"
The output will be a dataset with feature "a", "proportion rate", "interesting rate"


}
