% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kprot_cluster.R
\name{kprot_cluster}
\alias{kprot_cluster}
\title{Title}
\usage{
kprot_cluster(dat1, dat2, n = NULL)
}
\arguments{
\item{dat1}{data.frame}

\item{dat2}{data.frame}

\item{n}{integer}
}
\value{
plot
}
\description{
Title
}
