% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csv_skip_n.R
\name{read_all_csv_skip_n}
\alias{read_all_csv_skip_n}
\title{read_all_csv_skip_n}
\usage{
read_all_csv_skip_n(path, n = 2, pattern = "*.csv")
}
\arguments{
\item{path}{character}

\item{n}{integer}

\item{pattern}{character}
}
\value{
list
}
\description{
This function is used to read all csv in a folder
It return a list of all datasets.
examples: datalist <- read_all_csv_skip_n(path)
The first csv is datalist[[1]]
}
\examples{

}
