% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_plots.R
\name{spider_plots}
\alias{spider_plots}
\title{spider plots}
\usage{
spider_plots(good_data, feature, Time)
}
\arguments{
\item{good_data}{path}

\item{feature}{vector}

\item{Time}{character}
}
\value{
plot
}
\description{
This function will draw a spider_plots of all features
right now, it can only draw Ages & Time vs metrics
for example, spider_plots(good_data, "Age", "2014-Q1")
or spider_plots(good_data, "Age", "all") to draw all times automaticlly.
to be continue.
}
