% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggR_possible.R
\name{aggR_possible}
\alias{aggR_possible}
\title{aggR_possible}
\usage{
aggR_possible(dat, feature_names = names(dat))
}
\arguments{
\item{dat}{data.frame}

\item{feature_names}{vector}
}
\value{
filter_list list
}
\description{
Checks through a dataframe and a vector of features that can be deleted
outputs a list of list
each sublist contains a set of features that can be filtered out.
}
\examples{
load("../example_data/yummy.Rda")
dat_2

aggR_possible(dat_2, c("Colour", "Sweet_or_Salty", "Fruit"))


}
