% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggR_possible.R
\name{aggR_possible}
\alias{aggR_possible}
\title{aggR_possible}
\usage{
aggR_possible(data, features = names(data), keep = TRUE, all_symbol = "")
}
\arguments{
\item{data}{data.frame}

\item{features}{vector}

\item{keep}{logical}

\item{all_symbol}{character}
}
\value{
filter_list list
}
\description{
Checks through a dataframe and a vector of features that can be deleted
OR KEPT!!!!!
outputs a list of list
each sublist contains a set of features that can be filtered out.
To filter by keeping, set keep = TRUE (default)
To filter by columns to filter out, set keep = FALSE
}
\examples{
load("../example_data/yummy.Rda")
dat_2

#aggR_possible(dat_2, features =c("Colour", "Sweet_or_Salty", "Fruit"), keep = TRUE)


}
