% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRviz_join.R
\name{aggRviz_join}
\alias{aggRviz_join}
\title{Join 2 tables on the maximum overlapping set of features}
\usage{
aggRviz_join(x, y, key = c("measure", "rate"), col_2_ignore = NULL,
  all_symbol = "")
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}

\item{key}{vector}

\item{col_2_ignore}{vector}

\item{all_symbol}{character}
}
\value{
data.frame
}
\description{
This function will pick key column (such as "interesting rate" or proportion rate)
and join two lowest datasets by common columns.
}
\details{
THis is a new attempt, building on Fang's work

The key is a vector of key words that make something important, (and included.)
}
\examples{
load("../example_data/yummy.Rda")
aggRviz_join(dat_1, dat_2)

}
