% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_blanks.R
\name{filter_blanks}
\alias{filter_blanks}
\title{Filter aggregated data to the most granular level}
\usage{
filter_blanks(data, features = NULL, all_symbol = "")
}
\arguments{
\item{data}{data.frame}

\item{features}{vector}

\item{all_symbol}{character}
}
\value{
data.frame
}
\description{
Create a filtered dataframe with no blanks
Only on the features!!
}
\examples{
load("../example_data/yummy.Rda")
dat_2
filter_blanks(dat_2)

}
