% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_measures.R
\name{identify_measures}
\alias{identify_measures}
\title{identify_measures}
\usage{
identify_measures(data, key = c("measure", "rate"))
}
\arguments{
\item{data}{data.frame or vector}

\item{key}{vector}
}
\value{
vector
}
\description{
Enter a data.frame or vector, and this function will return a vector of measures
Optional:  add a vector of names to recognize as measures,
default is c("measure", "rate").
The function uses stringr to detect these regular expressions.
}
\examples{
load("../example_data/yummy.Rda")
names(dat_1)
identify_measures(dat_1)

}
