% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_measures.R
\name{identify_measures}
\alias{identify_measures}
\title{Title}
\usage{
identify_measures(data, key = c("measure", "rate"))
}
\arguments{
\item{data}{data.frame or vector}

\item{key}{vector}
}
\value{
vector
}
\description{
Title
}
